docs : pages-Chatting type 경로 변경
- 채팅페이지 type 분리
- import 순서 변경
