+++ src/pages/Chatting/ChatList/components/ChatListBox.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import ChatListLeft from './ChatListLeft';
 import ChatListRight from './ChatListRight';
-import { ChatListProp } from 'types/type';
+import { ChatListProp } from 'pages/Chatting/type';
 
 const ChatListBox = ({
   id,
