+++ src/pages/Chatting/ChatList/components/ChatListRight.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
-import ArrowRight from 'assets/svg/arrow-right.svg';
-import { ChatListProp } from 'types/type';
 import { useState } from 'react';
 import ChatModal from 'pages/Chatting/ChatModal';
+import { ChatListProp } from 'pages/Chatting/type';
+import ArrowRight from 'assets/svg/arrow-right.svg';
 
 const ChatListRight = ({
   id,
