+++ src/pages/Chatting/type.ts
@@ -1,0 +1,17 @@
+export interface ChatListProp {
+  id?: number;
+  Image?: any;
+  title?: string;
+  description?: string;
+  modalDescription?: string;
+  type?: string;
+}
+
+export interface MessagesProps {
+  user?: string;
+  text?: string;
+  time?: string;
+  mbti?: string;
+  nickname?: string;
+  message;
+}
